+++ migrations/20260312170557_create_group_profiles_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS group_profiles (
     group_id TEXT PRIMARY KEY,
     encrypted_blob BLOB NOT NULL,
-    last_updated BIGINT NOT NULL,
+    last_updated INTEGER NOT NULL,
     version INTEGER DEFAULT 1
 );
